backend-temp/requirements.txt
dj_database_url==0.5.0
requests
Django==2.2.1  
gunicorn==19.9.0
psycopg2==2.8.3
pytz==2019.1
whitenoise==4.1.2
djangorestframework==3.9.4
django-cors-headers==3.0.2
django-jet==1.0.8  
markdown
django-filter
elasticsearch